+++ appc/conn25.go
@@ -4,57 +4,16 @@
 package appc
 
 import (
-	"cmp"
 	"fmt"
-	"slices"
 	"strings"
 
-	"tailscale.com/ipn/ipnext"
 	"tailscale.com/tailcfg"
 	"tailscale.com/tailcfg/nodecap"
 	"tailscale.com/types/appctype"
 	"tailscale.com/types/dnstype"
-	"tailscale.com/util/set"
 )
 
 const AppConnectorsExperimentalAttrName = "tailscale.com/app-connectors-experimental"
-
-func isPeerEligibleConnector(peer tailcfg.NodeView) bool {
-	if !peer.Valid() || !peer.Hostinfo().Valid() {
-		return false
-	}
-	isConn, _ := peer.Hostinfo().AppConnector().Get()
-	return isConn
-}
-
-func sortByPreference(ns []tailcfg.NodeView) {
-	// The ordering of the nodes is semantic (callers use the first node they can
-	// get a peer api url for). We don't (currently 2026-02-27) have any
-	// preference over which node is chosen as long as it's consistent.  In the
-	// future we anticipate integrating with traffic steering.
-	slices.SortFunc(ns, func(a, b tailcfg.NodeView) int {
-		return cmp.Compare(a.ID(), b.ID())
-	})
-}
-
-// PickConnector returns peers the backend knows about that match the app, in order of preference to use as
-// a connector.
-func PickConnector(nb ipnext.NodeBackend, app appctype.Conn25Attr) []tailcfg.NodeView {
-	appTagsSet := set.SetOf(app.Connectors)
-	matches := nb.AppendMatchingPeers(nil, func(n tailcfg.NodeView) bool {
-		if !isPeerEligibleConnector(n) {
-			return false
-		}
-		for _, t := range n.Tags().All() {
-			if appTagsSet.Contains(t) {
-				return true
-			}
-		}
-		return false
-	})
-	sortByPreference(matches)
-	return matches
-}
 
 // DNSAddrScheme is the custom URI scheme used for conn25-managed split DNS
 // entries to determine the destination at query time rather than configuration
